adapter: Merge two cli flags into DeploymentMode enum
This commit merges the `--standalone` and `--embedded-readers` cli flags
and their env vars into the new DeploymentMode enum (flag and env var).
This enum helps disambiguate the behavior of the previous two boolean flags,
and enables us to make standalone the default mode (which was not possible
given the two booleans and how defaults work with clap).

Further, this commit updates all uses of the previous flags and env vars.

Fixes: REA-3585
Release-Core-Note: Merging `--standalone` and `--embedded-readers`
  flags into new `--deployment-mode` enum.
Change-Id: I9b1a45f32c0ceb9df69de08e6df081f426d8a0df