adapter/server: change default deployment name
The `--deployment` states the id of the current readyset deployment.
This id must be shared by all the nodes in a deployment (via the
authority), as well as the name of a directory on the local disk
(a subdirectory of `--storage_dir`). The existing default name
for the deployment is `tmp-readyset`. In order to avoid any confusion
an operator might have, let's rename this to something less
"temporary".

Fixes: REA-3585
Release-Core-Note: Change the default ReadySet deployment name
  from `tmp-readyset` to `readyset.db`.
Change-Id: I7e80a645a96a2dc2a31799d6b6a96d3005d2f6ab